+++ packages/blorktools/src/asset_debugger/util/rig/rig-factory.js
@@ -38,1022 +38,9 @@ import {
     resetBones
   } from './bone-kinematics.js';
 import { createAxisIndicator } from '../../axis-indicator/axis-indicator';
-
-/**
- * Create joint labels for all joints in the scene
- * @param {Object} scene - The Three.js scene
- */
-export function createJointLabels(scene) {
-    console.log('Creating joint labels...');
-    
-    // Remove any existing labels first
-    clearJointLabels(scene);
-    
-    // Create a group to hold all labels
-    setJointLabelsGroup("JointLabels", scene);
-    
-    // Keep track of the labels created
-    const labelCount = {total: 0, added: 0};
-    
-    // Track positions where we've already created labels to prevent duplicates
-    const labelPositions = [];
-    const positionTolerance = 0.001; // Tolerance for considering positions as identical
-    
-    // Helper function to check if a position already has a label
-    const hasLabelAtPosition = (position) => {
-        return labelPositions.some(pos => position.distanceTo(pos) < positionTolerance);
-    };
-    
-    // Find all bone meshes
-    boneVisualsGroup.traverse((object) => {
-        if (object.userData && object.userData.bonePart === 'cap') {
-            labelCount.total++;
-            
-            // Get world position of this joint
-            const worldPos = new THREE.Vector3();
-            object.getWorldPosition(worldPos);
-            
-            // Check if we already have a label at this position
-            if (hasLabelAtPosition(worldPos)) {
-                console.log('Skipping duplicate joint label at position:', worldPos);
-                return;
-            }
-            
-            // Determine which bone name to use
-            let boneName = "";
-            if (object.parent && object.parent.userData) {
-                if (object.position.y > 0 && object.parent.userData.childBone) {
-                    // Top sphere - use child bone name
-                    boneName = object.parent.userData.childBone.name;
-                } else if (object.position.y === 0 && object.parent.userData.parentBone) {
-                    // Bottom sphere - use parent bone name
-                    boneName = object.parent.userData.parentBone.name;
-                }
-            }
-            
-            if (boneName) {
-                // Create a label for this joint
-                const label = createSimpleLabel(boneName, object, scene);
-                if (label) {
-                    const jointLabelGroup = getJointLabelGroup();
-                    if (jointLabelGroup) {
-                        jointLabelGroup.add(label);
-                        labelCount.added++;
-                        
-                        // Record this position as having a label
-                        labelPositions.push(worldPos);
-                    }
-                }
-            }
-        }
-    });
-    
-    console.log(`Created ${labelCount.added} labels out of ${labelCount.total} joint spheres found`);
-    return scene.getObjectByName("JointLabels");
-}
-
-/**
- * Create bone labels for all bone segments in the scene
- * @param {Object} scene - The Three.js scene
- */
-export function createBoneLabels(scene) {
-    console.log('Creating bone labels...');
-    
-    // Remove any existing bone labels
-    clearBoneLabels(scene);
-    
-    // Create a group to hold all bone labels
-    setBoneLabelsGroup("BoneLabels", scene);
-    
-    // Keep track of the labels created
-    const labelCount = {total: 0, added: 0};
-    
-    // Track bone connections that already have labels to prevent duplicates
-    const labeledBoneConnections = new Set();
-    
-    // Find all bone groups
-    boneVisualsGroup.children.forEach((boneGroup) => {
-        // Skip if not a bone group
-        if (!boneGroup.userData || !boneGroup.userData.isVisualBone) {
-            return;
-        }
-        
-        labelCount.total++;
-        
-        // Get bone name information from the bone group
-        let boneName = "";
-        if (boneGroup.userData.parentBone && boneGroup.userData.childBone) {
-            // Create a connection identifier (both directions to handle differently ordered pairs)
-            const parentID = boneGroup.userData.parentBone.id || boneGroup.userData.parentBone.uuid;
-            const childID = boneGroup.userData.childBone.id || boneGroup.userData.childBone.uuid;
-            const connectionID = `${parentID}_${childID}`;
-            const reverseConnectionID = `${childID}_${parentID}`;
-            
-            // Skip if we already created a label for this connection
-            if (labeledBoneConnections.has(connectionID) || labeledBoneConnections.has(reverseConnectionID)) {
-                console.log('Skipping duplicate bone label for connection:', 
-                           `${boneGroup.userData.parentBone.name} → ${boneGroup.userData.childBone.name}`);
-                return;
-            }
-            
-            // Create a name that indicates the connection
-            boneName = `${boneGroup.userData.parentBone.name} → ${boneGroup.userData.childBone.name}`;
-            
-            // Create a label for this bone
-            const label = createSimpleLabel(boneName, boneGroup, scene, true); // Pass true to indicate it's a bone label
-            if (label) {
-                const boneLabelGroup = getBoneLabelGroup();
-                if (boneLabelGroup) {
-                    boneLabelGroup.add(label);
-                    labelCount.added++;
-                    
-                    // Mark this connection as labeled
-                    labeledBoneConnections.add(connectionID);
-                    
-                    // Calculate the midpoint between parent and child bones immediately
-                    const parentPos = new THREE.Vector3();
-                    const childPos = new THREE.Vector3();
-                    
-                    boneGroup.userData.parentBone.getWorldPosition(parentPos);
-                    boneGroup.userData.childBone.getWorldPosition(childPos);
-                    
-                    // Calculate the middle point
-                    const midPoint = new THREE.Vector3().addVectors(parentPos, childPos).multiplyScalar(0.5);
-                    
-                    // Position at the middle point immediately
-                    label.position.copy(midPoint);
-                    
-                    // Position the label at the middle of the bone
-                    if (label.userData.updatePosition) {
-                        label.userData.updatePosition();
-                    }
-                }
-            }
-        }
-    });
-    
-    console.log(`Created ${labelCount.added} labels out of ${labelCount.total} bone segments found`);
-    return scene.getObjectByName("BoneLabels");
-}
-
-/**
- * Create a simple text label as a sprite
- * @param {string} text - Text to display
- * @param {Object} joint - Joint object to attach to
- * @param {Object} scene - Three.js scene
- * @param {boolean} isBoneLabel - Whether this is a bone label (true) or joint label (false)
- * @returns {Object} The created label sprite
- */
-function createSimpleLabel(text, joint, scene, isBoneLabel = false) {
-    console.log(`Creating label for ${isBoneLabel ? 'bone' : 'joint'}: ${text}`);
-
-    // Create a canvas for the label
-    const canvas = document.createElement('canvas');
-    const ctx = canvas.getContext('2d');
-    
-    // Set canvas size
-    canvas.width = 256;
-    // Increase height for bone labels to accommodate multiple lines
-    canvas.height = isBoneLabel ? 96 : 64;
-    
-    // Different styling for bone vs joint labels - ONLY VISUAL DIFFERENCES HERE
-    const labelConfig = {
-        bgColor: isBoneLabel ? 'rgba(30, 136, 229, 0.8)' : 'rgba(76, 175, 80, 0.8)',
-        gradientStops: isBoneLabel ? 
-            [{ pos: 0, color: 'rgba(40, 40, 80, 0.4)' }, { pos: 1, color: 'rgba(20, 20, 40, 0.6)' }] : 
-            [{ pos: 0, color: 'rgba(30, 60, 30, 0.4)' }, { pos: 1, color: 'rgba(10, 30, 10, 0.6)' }],
-        borderColor: isBoneLabel ? '#88AAFF' : '#AAFFAA',
-        scrollThumbColor: isBoneLabel ? '#88AAFF' : '#AAFFAA',
-        headerText: isBoneLabel ? 'BONE' : 'JOINT'
-    };
-    
-    // Background fill
-    ctx.fillStyle = labelConfig.bgColor;
-    ctx.fillRect(0, 0, canvas.width, canvas.height);
-    
-    // Add gradient overlay
-    const gradient = ctx.createLinearGradient(0, 0, 0, canvas.height);
-    labelConfig.gradientStops.forEach(stop => {
-        gradient.addColorStop(stop.pos, stop.color);
-    });
-    ctx.fillStyle = gradient;
-    ctx.fillRect(0, 0, canvas.width, canvas.height);
-    
-    // Border
-    ctx.strokeStyle = labelConfig.borderColor;
-    ctx.lineWidth = 2;
-    ctx.strokeRect(0, 0, canvas.width, canvas.height);
-    
-    // Header line
-    ctx.fillStyle = 'rgba(0, 0, 0, 0.3)';
-    ctx.fillRect(0, 0, canvas.width, 18);
-    
-    // Header text
-    ctx.font = 'bold 12px monospace';
-    ctx.fillStyle = '#DDDDDD';
-    ctx.textAlign = 'left';
-    ctx.textBaseline = 'middle';
-    ctx.fillText(labelConfig.headerText, 6, 9);
-    
-    // Main text
-    ctx.font = 'bold 16px monospace';
-    ctx.fillStyle = 'white';
-    ctx.textAlign = 'center';
-    ctx.textBaseline = 'middle';
-    
-    // For bone labels, format as multiple lines if it contains an arrow
-    if (isBoneLabel && text.includes('→')) {
-        const parts = text.split('→');
-        const parent = parts[0].trim();
-        const child = parts[1].trim();
-        
-        // Calculate max width to check if parts need scrolling
-        const parentWidth = ctx.measureText(parent).width;
-        const childWidth = ctx.measureText(child).width;
-        const maxWidth = canvas.width - 20; // 10px padding on each side
-        
-        // First line: parent
-        let parentText = parent;
-        if (parentWidth > maxWidth) {
-            parentText = parent.length > 25 ? parent.substring(0, 22) + '...' : parent;
-        }
-        ctx.fillText(parentText, canvas.width / 2, 35);
-        
-        // Arrow line
-        ctx.fillText('↓', canvas.width / 2, 55);
-        
-        // Third line: child
-        let childText = child;
-        if (childWidth > maxWidth) {
-            childText = child.length > 25 ? child.substring(0, 22) + '...' : child;
-        }
-        ctx.fillText(childText, canvas.width / 2, 75);
-        
-        // If either part was truncated, add a scroll indicator
-        if (parentWidth > maxWidth || childWidth > maxWidth) {
-            // Draw scroll indicator
-            ctx.fillStyle = 'rgba(255, 255, 255, 0.4)';
-            ctx.fillRect(10, canvas.height - 8, canvas.width - 20, 4);
-            
-            // Draw scrollbar thumb
-            const contentWidth = Math.max(parentWidth, childWidth);
-            const thumbWidth = Math.max(30, (maxWidth / contentWidth) * (canvas.width - 20));
-            ctx.fillStyle = labelConfig.scrollThumbColor;
-            ctx.fillRect(10, canvas.height - 8, thumbWidth, 4);
-        }
-    } else {
-        // Regular single-line text handling (for joint labels or non-arrow bone labels)
-        // Calculate text width to check if it needs scrolling
-        const textMetrics = ctx.measureText(text);
-        const textWidth = textMetrics.width;
-        const maxWidth = canvas.width - 20; // 10px padding on each side
-        
-        // If text is too long, implement scrolling behavior
-        if (textWidth > maxWidth) {
-            // Indicate text is scrollable with ellipsis
-            const displayText = text.length > 25 ? text.substring(0, 22) + '...' : text;
-            ctx.fillText(displayText, canvas.width / 2, canvas.height / 2 + 5);
-            
-            // Draw scroll indicator
-            ctx.fillStyle = 'rgba(255, 255, 255, 0.4)';
-            ctx.fillRect(10, canvas.height - 8, canvas.width - 20, 4);
-            
-            // Draw scrollbar thumb
-            const thumbWidth = Math.max(30, (maxWidth / textWidth) * (canvas.width - 20));
-            ctx.fillStyle = labelConfig.scrollThumbColor;
-            ctx.fillRect(10, canvas.height - 8, thumbWidth, 4);
-        } else {
-            // Text fits, just display it
-            ctx.fillText(text, canvas.width / 2, canvas.height / 2 + 5);
-        }
-    }
-    
-    // Create sprite material
-    const texture = new THREE.CanvasTexture(canvas);
-    const material = new THREE.SpriteMaterial({
-        map: texture,
-        transparent: true,
-        depthTest: false,
-        depthWrite: false
-    });
-    
-    // Create sprite
-    const sprite = new THREE.Sprite(material);
-    sprite.userData.isBoneLabel = isBoneLabel;
-    sprite.userData.isJointLabel = !isBoneLabel;
-    sprite.userData.targetJoint = joint;
-    sprite.userData.labelText = text; // Store original text
-    
-    // Add hover handling for label headers
-    sprite.userData.isInteractive = true;
-    sprite.userData.isLabelHeader = true;
-    sprite.userData.headerHeight = 18; // Height of the header section in pixels
-    sprite.userData.canvasHeight = canvas.height;
-    sprite.userData.canvasWidth = canvas.width;
-    
-    // Add mouse event handling to prevent propagation when over the header
-    sprite.userData.onMouseMove = (event) => {
-        // Check if mouse is over the header area
-        if (sprite.userData.isMouseOverHeader) {
-            // Prevent event propagation to stop camera controls
-            event.stopPropagation();
-        }
-    };
-    
-    // This function calculates if mouse is over the header or the content area
-    sprite.userData.checkHeaderHover = (mousePos) => {
-        if (!mousePos) return false;
-        
-        // The sprite's coordinates go from -0.5 to 0.5 in both dimensions
-        // So we need to convert from local sprite coordinates to canvas coordinates
-        
-        // Calculate the overall dimensions in world units
-        const totalHeight = sprite.scale.y;
-        
-        // Convert mousePos.y from local coordinates (-0.5 to 0.5) to normalized height (0 to 1)
-        // In THREE.js, sprite local coords have origin at center, Y+ is up
-        // -0.5 is bottom, 0.5 is top, so we need to flip and shift
-        const normalizedY = 0.5 - mousePos.y; // Convert to 0 at top, 1 at bottom
-        
-        // Now convert to canvas pixel coordinates
-        const canvasY = normalizedY * canvas.height;
-        
-        // Check if we're in the header region (top 18px of the canvas)
-        const isOverHeader = canvasY >= 0 && canvasY <= sprite.userData.headerHeight;
-        
-        return isOverHeader;
-    };
-    
-    // Set initial position safely - updateLabelPosition now has defensive coding
-    updateLabelPosition(sprite, joint);
-    
-    // FIXED SIZE FOR ALL LABELS - no dependency on underlying object geometry
-    const fixedScale = 0.3;
-    // Adjust height for bone labels to accommodate multiple lines
-    const heightFactor = isBoneLabel ? 0.4 : 0.25;
-    sprite.scale.set(fixedScale, fixedScale * heightFactor, 1);
-    
-    // Set initial visibility based on appropriate option
-    sprite.visible = isBoneLabel ? rigOptions.showBoneLabels : rigOptions.showJointLabels;
-    if (!sprite.visible) {
-        console.log(`Label for ${text} is initially hidden`);
-    }
-    
-    // Set up the update function - customize for bone labels
-    sprite.userData.updatePosition = () => {
-        // For bone labels, we want to position them in the middle of the bone
-        if (isBoneLabel) {
-            // Get the bone group which should have parent and child bone references
-            const boneGroup = joint.userData && joint.userData.isVisualBone ? 
-                joint : (joint.parent && joint.parent.userData && joint.parent.userData.isVisualBone ? 
-                    joint.parent : null);
-            
-            if (boneGroup && boneGroup.userData && 
-                boneGroup.userData.parentBone && boneGroup.userData.childBone) {
-                
-                // Get world positions of parent and child bones
-                const parentPos = new THREE.Vector3();
-                const childPos = new THREE.Vector3();
-                
-                boneGroup.userData.parentBone.getWorldPosition(parentPos);
-                boneGroup.userData.childBone.getWorldPosition(childPos);
-                
-                // Calculate the middle point
-                const midPoint = new THREE.Vector3().addVectors(parentPos, childPos).multiplyScalar(0.5);
-                
-                // Position exactly at the midpoint without any offset
-                sprite.position.copy(midPoint);
-            } else {
-                // Fallback for when we can't find the proper bone references
-                const worldPos = new THREE.Vector3();
-                joint.getWorldPosition(worldPos);
-                sprite.position.copy(worldPos);
-            }
-        } else {
-            // For joint labels, use the standard update function
-            updateLabelPosition(sprite, joint);
-        }
-    };
-    
-    // Initialize the position
-    if (sprite.userData.updatePosition) {
-        sprite.userData.updatePosition();
-    }
-    
-    // Make sure the sprite renders on top
-    sprite.renderOrder = isBoneLabel ? 990 : 1000; // Joint labels show on top of bone labels
-    
-    return sprite;
-}
-
-/**
- * Parse joint constraint data from node extras or userData
- * @param {Object} node - Bone/joint node to examine
- * @returns {Object|null} - Constraint data if found, null otherwise
- */
-export function parseJointConstraints(node) {
-    console.log(`Checking for constraints on joint: ${node.name}`);
-    
-    // Initialize constraint object
-    let constraints = null;
-    
-    // Check for constraints in userData
-    if (node.userData) {
-        // Look for Blender constraint data in userData
-        if (node.userData.constraints || node.userData.boneConstraints) {
-            constraints = node.userData.constraints || node.userData.boneConstraints;
-            console.log(`Found explicit constraints in userData for ${node.name}:`, constraints);
-            return constraints;
-        }
-        
-        // Check for limit_rotation type constraints
-        if (node.userData.limitRotation || node.userData.rotationLimits) {
-            constraints = {
-                type: 'limitRotation',
-                limits: node.userData.limitRotation || node.userData.rotationLimits
-            };
-            console.log(`Found rotation limits for ${node.name}:`, constraints);
-            return constraints;
-        }
-    }
-    
-    // Look for GLTF extras (where custom data is often stored)
-    if (node.extras) {
-        if (node.extras.constraints || node.extras.jointType) {
-            constraints = node.extras.constraints || { type: node.extras.jointType };
-            console.log(`Found constraints in GLTF extras for ${node.name}:`, constraints);
-            return constraints;
-        }
-    }
-    
-    // Infer constraints from bone properties
-    // Check if bone has locked rotation axes
-    if (node.rotation && node.userData.initialRotation) {
-        // Look for zero initial rotations that might indicate locked axes
-        const lockedAxes = [];
-        const epsilon = 0.0001; // Small threshold for floating point comparison
-        
-        if (Math.abs(node.userData.initialRotation.x) < epsilon) lockedAxes.push('x');
-        if (Math.abs(node.userData.initialRotation.y) < epsilon) lockedAxes.push('y');
-        if (Math.abs(node.userData.initialRotation.z) < epsilon) lockedAxes.push('z');
-        
-        // If we have 2+ locked axes, this might be a hinge joint
-        if (lockedAxes.length >= 2) {
-            const freeAxis = ['x', 'y', 'z'].find(axis => !lockedAxes.includes(axis));
-            if (freeAxis) {
-                constraints = {
-                    type: 'hinge',
-                    axis: freeAxis
-                };
-                console.log(`Inferred hinge joint for ${node.name} on ${freeAxis} axis from locked rotations`);
-                return constraints;
-            }
-        }
-        
-        // If all axes are locked, this might be a fixed joint
-        if (lockedAxes.length === 3) {
-            constraints = {
-                type: 'fixed'
-            };
-            console.log(`Inferred fixed joint for ${node.name} from locked rotations on all axes`);
-            return constraints;
-        }
-    }
-    
-    // Check naming conventions that might indicate constraint types
-    const lowerName = node.name.toLowerCase();
-    if (lowerName.includes('fixed') || lowerName.includes('rigid')) {
-        constraints = {
-            type: 'fixed'
-        };
-        console.log(`Inferred fixed joint for ${node.name} from naming convention`);
-        return constraints;
-    }
-    
-    if (lowerName.includes('hinge') || lowerName.includes('elbow') || lowerName.includes('knee')) {
-        // Default to Y axis for hinge if not specified
-        constraints = {
-            type: 'hinge',
-            axis: lowerName.includes('_x') ? 'x' : (lowerName.includes('_z') ? 'z' : 'y')
-        };
-        console.log(`Inferred hinge joint for ${node.name} from naming convention`);
-        return constraints;
-    }
-    
-    if (lowerName.includes('spring') || lowerName.includes('bounce')) {
-        constraints = {
-            type: 'spring',
-            stiffness: 50,  // Default stiffness
-            damping: 5      // Default damping
-        };
-        console.log(`Inferred spring joint for ${node.name} from naming convention`);
-        return constraints;
-    }
-    
-    // No constraints found
-    return null;
-}
-
-/**
- * Apply joint constraints to a bone
- * @param {Object} bone - The bone to apply constraints to
- * @param {Object} constraints - Constraint data
- */
-export function applyJointConstraints(bone, constraints) {
-    if (!bone || !constraints) return;
-    
-    console.log(`Applying ${constraints.type} constraint to ${bone.name}`);
-    
-    // Store constraint data in bone userData for reference
-    bone.userData.constraints = constraints;
-    
-    // Check if we should preserve the current position
-    const preserveCurrentPose = constraints.preservePosition && 
-                               constraints.currentPosition && 
-                               constraints.currentQuaternion;
-    
-    // If we're preserving the current position, store it before applying constraints
-    if (preserveCurrentPose) {
-        console.log(`Preserving current pose for ${bone.name}`);
-        bone.userData.preservedPosition = constraints.currentPosition.clone();
-        bone.userData.preservedQuaternion = constraints.currentQuaternion.clone();
-    }
-    
-    // Apply different constraint types
-    switch (constraints.type) {
-        case 'fixed':
-            // Fixed joints don't allow rotation - enforce this by marking as locked
-            bone.userData.isLocked = true;
-            
-            // If we're preserving the current pose, store it as the fixed position
-            if (preserveCurrentPose) {
-                // For fixed constraints, we want to maintain the current world position
-                bone.userData.fixedPosition = constraints.currentPosition.clone();
-                bone.userData.fixedQuaternion = constraints.currentQuaternion.clone();
-                console.log(`Applied fixed constraint to ${bone.name} at current position - joint will be locked`);
-            } else {
-                console.log(`Applied fixed constraint to ${bone.name} - joint will be locked`);
-            }
-            break;
-            
-        case 'hinge':
-            // Hinge joints only allow rotation on one axis
-            bone.userData.hinge = {
-                axis: constraints.axis || 'y',
-                min: constraints.min !== undefined ? constraints.min : -Math.PI/2,
-                max: constraints.max !== undefined ? constraints.max : Math.PI/2
-            };
-            
-            // If preserving position, store current rotation as the initial state
-            if (preserveCurrentPose) {
-                bone.userData.initialWorldQuaternion = constraints.currentQuaternion.clone();
-            }
-            
-            console.log(`Applied hinge constraint to ${bone.name} on ${bone.userData.hinge.axis} axis with limits [${bone.userData.hinge.min}, ${bone.userData.hinge.max}]`);
-            break;
-            
-        case 'limitRotation':
-            // Limit rotation within specific ranges for each axis
-            bone.userData.rotationLimits = {
-                x: constraints.limits?.x || { min: -Math.PI/4, max: Math.PI/4 },
-                y: constraints.limits?.y || { min: -Math.PI/4, max: Math.PI/4 },
-                z: constraints.limits?.z || { min: -Math.PI/4, max: Math.PI/4 }
-            };
-            
-            // If preserving position, store current rotation
-            if (preserveCurrentPose) {
-                bone.userData.initialWorldQuaternion = constraints.currentQuaternion.clone();
-            }
-            
-            console.log(`Applied rotation limits to ${bone.name}:`, bone.userData.rotationLimits);
-            break;
-            
-        case 'spring':
-            // Spring joints try to return to their rest position
-            bone.userData.spring = {
-                stiffness: constraints.stiffness || 50,
-                damping: constraints.damping || 5,
-                gravity: constraints.gravity || 1.0,
-                // Store the current pose as the rest position without applying immediate forces
-                restPosition: preserveCurrentPose 
-                    ? new THREE.Vector3().copy(constraints.currentPosition) 
-                    : new THREE.Vector3().copy(bone.position),
-                restRotation: preserveCurrentPose
-                    ? new THREE.Euler().setFromQuaternion(constraints.currentQuaternion)
-                    : new THREE.Euler(
-                        bone.rotation.x,
-                        bone.rotation.y,
-                        bone.rotation.z,
-                        bone.rotation.order
-                    ),
-                // Add velocity tracking for proper spring physics
-                velocity: new THREE.Vector3(0, 0, 0),
-                lastPosition: null, // Will be set during first update
-                lastTime: Date.now()
-            };
-            console.log(`Applied spring constraint to ${bone.name} with stiffness ${bone.userData.spring.stiffness} - rest pose preserved`);
-            break;
-            
-        case 'none':
-        default:
-            // No constraints (default) - unrestricted rotation
-            // No need to do anything special here, allow free movement
-            break;
-    }
-    
-    // Add a custom update function to enforce constraints during animation
-    const originalUpdateMatrix = bone.updateMatrix;
-    bone.updateMatrix = function() {
-        // Apply constraints to rotation before updating matrix
-        if (this.userData.constraints) {
-            enforceJointConstraints(this);
-        }
-        // Call the original update function
-        originalUpdateMatrix.call(this);
-    };
-    
-    // If we're preserving position, apply it now
-    if (preserveCurrentPose) {
-        // For fixed constraints, directly set the world position and quaternion
-        if (constraints.type === 'fixed' && bone.parent) {
-            // We need to convert world position back to local
-            const worldPos = bone.userData.preservedPosition;
-            const worldQuat = bone.userData.preservedQuaternion;
-            
-            // Convert to parent space
-            const parentWorldInverse = new THREE.Matrix4().invert(bone.parent.matrixWorld);
-            const localPos = worldPos.clone().applyMatrix4(parentWorldInverse);
-            
-            const parentQuatInverse = new THREE.Quaternion().copy(bone.parent.quaternion).invert();
-            const localQuat = worldQuat.clone().multiply(parentQuatInverse);
-            
-            // Apply the local position and rotation
-            bone.position.copy(localPos);
-            bone.quaternion.copy(localQuat);
-        }
-        
-        // Update the bone matrix to reflect these changes
-        bone.updateMatrix();
-    }
-}
-
-/**
- * Enforce joint constraints on a bone
- * @param {Object} bone - The bone to enforce constraints on
- */
-export function enforceJointConstraints(bone) {
-    if (!bone || !bone.userData.constraints) return;
-    
-    const constraints = bone.userData.constraints;
-    
-    switch (constraints.type) {
-        case 'fixed':
-            // Fixed joints - maintain position and orientation
-            if (bone.userData.fixedPosition && bone.userData.fixedQuaternion) {
-                // Use the stored fixed position and orientation (world coordinates)
-                // We need to convert back to local space relative to the parent
-                if (bone.parent) {
-                    // Get parent world inverse
-                    const parentWorldInverse = new THREE.Matrix4().copy(bone.parent.matrixWorld).invert();
-                    const localPos = bone.userData.fixedPosition.clone().applyMatrix4(parentWorldInverse);
-                    
-                    // Get parent quaternion inverse
-                    const parentQuatInverse = new THREE.Quaternion().copy(bone.parent.quaternion).invert();
-                    const localQuat = new THREE.Quaternion().copy(bone.userData.fixedQuaternion)
-                        .premultiply(parentQuatInverse);
-                    
-                    // Set local position and rotation
-                    bone.position.copy(localPos);
-                    bone.quaternion.copy(localQuat);
-                }
-            }
-            // If no fixed position is stored, use initial rotation
-            else if (bone.userData.initialRotation) {
-                bone.rotation.set(
-                    bone.userData.initialRotation.x,
-                    bone.userData.initialRotation.y,
-                    bone.userData.initialRotation.z
-                );
-                bone.rotation.order = bone.userData.initialRotation.order;
-            }
-            break;
-            
-        case 'hinge':
-            // Hinge joints - only allow rotation on one axis, reset others
-            if (bone.userData.hinge) {
-                const hinge = bone.userData.hinge;
-                const initial = bone.userData.initialRotation || { x: 0, y: 0, z: 0 };
-                
-                // Reset rotation on locked axes
-                if (hinge.axis !== 'x') bone.rotation.x = initial.x;
-                if (hinge.axis !== 'y') bone.rotation.y = initial.y;
-                if (hinge.axis !== 'z') bone.rotation.z = initial.z;
-                
-                // Clamp rotation on the free axis
-                if (hinge.axis === 'x') {
-                    bone.rotation.x = Math.max(hinge.min, Math.min(hinge.max, bone.rotation.x));
-                } else if (hinge.axis === 'y') {
-                    bone.rotation.y = Math.max(hinge.min, Math.min(hinge.max, bone.rotation.y));
-                } else if (hinge.axis === 'z') {
-                    bone.rotation.z = Math.max(hinge.min, Math.min(hinge.max, bone.rotation.z));
-                }
-            }
-            break;
-            
-        case 'limitRotation':
-            // Apply rotation limits on each axis
-            if (bone.userData.rotationLimits) {
-                const limits = bone.userData.rotationLimits;
-                
-                if (limits.x) {
-                    bone.rotation.x = Math.max(limits.x.min, Math.min(limits.x.max, bone.rotation.x));
-                }
-                if (limits.y) {
-                    bone.rotation.y = Math.max(limits.y.min, Math.min(limits.y.max, bone.rotation.y));
-                }
-                if (limits.z) {
-                    bone.rotation.z = Math.max(limits.z.min, Math.min(limits.z.max, bone.rotation.z));
-                }
-            }
-            break;
-            
-        case 'spring':
-            // Ragdoll-like spring implementation that respects hierarchy
-            if (bone.userData.spring) {
-                const spring = bone.userData.spring;
-                
-                // Initialize tracking if this is the first update
-                const now = Date.now();
-                if (!spring.lastPosition) {
-                    spring.lastPosition = new THREE.Vector3().copy(bone.position);
-                    spring.lastTime = now;
-                    spring.velocityX = 0;
-                    spring.velocityY = 0;
-                    spring.velocityZ = 0;
-                    spring.lastRotation = new THREE.Euler().copy(bone.rotation);
-                    break;
-                }
-                
-                // Calculate time delta with safeguards
-                const deltaTime = Math.min((now - spring.lastTime) / 1000, 0.05); // Cap at 50ms for stability
-                if (deltaTime <= 0) break;
-                
-                // Store hierarchy information
-                if (!spring.hierarchyFactor) {
-                    // Set influence based on depth in hierarchy (1.0 for root, decreases for children)
-                    let hierarchyDepth = 0;
-                    let parent = bone.parent;
-                    while (parent && parent.isBone) {
-                        hierarchyDepth++;
-                        parent = parent.parent;
-                    }
-                    // Deeper bones receive less influence to prevent cascading oscillations
-                    spring.hierarchyFactor = Math.max(0.2, 1.0 / (hierarchyDepth + 1));
-                }
-                
-                // Calculate rotational differences from rest pose
-                const diffX = spring.restRotation.x - bone.rotation.x;
-                const diffY = spring.restRotation.y - bone.rotation.y;
-                const diffZ = spring.restRotation.z - bone.rotation.z;
-                
-                // Calculate rotational velocity (change since last frame)
-                const rotVelX = (bone.rotation.x - spring.lastRotation.x) / deltaTime;
-                const rotVelY = (bone.rotation.y - spring.lastRotation.y) / deltaTime;
-                const rotVelZ = (bone.rotation.z - spring.lastRotation.z) / deltaTime;
-                
-                // Proper spring forces adjusted by hierarchy factor and mass (simulated by bone size)
-                const boneSize = bone.scale.length() || 1;
-                const massInfluence = 1.0 / (boneSize * 0.5 + 0.5); // Larger bones have more mass
-                
-                // Apply hierarchy and mass adjustments to spring calculations
-                const adjustedStiffness = spring.stiffness * spring.hierarchyFactor * massInfluence;
-                const adjustedDamping = spring.damping * (1 + (1 - spring.hierarchyFactor) * 2); // More damping for deeper bones
-                
-                // Calculate spring force with adjusted parameters
-                const springForceX = diffX * adjustedStiffness;
-                const springForceY = diffY * adjustedStiffness;
-                const springForceZ = diffZ * adjustedStiffness;
-                
-                // Apply damping force proportional to current velocity
-                const dampingForceX = -rotVelX * adjustedDamping;
-                const dampingForceY = -rotVelY * adjustedDamping;
-                const dampingForceZ = -rotVelZ * adjustedDamping;
-                
-                // Combine forces
-                const totalForceX = springForceX + dampingForceX;
-                const totalForceY = springForceY + dampingForceY;
-                const totalForceZ = springForceZ + dampingForceZ;
-                
-                // Update velocities with forces
-                spring.velocityX += totalForceX * deltaTime;
-                spring.velocityY += totalForceY * deltaTime;
-                spring.velocityZ += totalForceZ * deltaTime;
-                
-                // Apply absolute velocity decay to ensure the motion eventually stops
-                // This decay factor is separate from the damping which only affects the instantaneous force
-                const velocityDecayFactor = Math.max(0, 1.0 - (spring.damping * 0.01 + 0.05) * deltaTime);
-                spring.velocityX *= velocityDecayFactor;
-                spring.velocityY *= velocityDecayFactor;
-                spring.velocityZ *= velocityDecayFactor;
-                
-                // Apply velocity threshold to stop tiny movements
-                const velocityThreshold = 0.001;
-                if (Math.abs(spring.velocityX) < velocityThreshold) spring.velocityX = 0;
-                if (Math.abs(spring.velocityY) < velocityThreshold) spring.velocityY = 0;
-                if (Math.abs(spring.velocityZ) < velocityThreshold) spring.velocityZ = 0;
-                
-                // Apply velocity limits to prevent extreme oscillations
-                const maxVelocity = 15.0;
-                spring.velocityX = Math.max(-maxVelocity, Math.min(maxVelocity, spring.velocityX));
-                spring.velocityY = Math.max(-maxVelocity, Math.min(maxVelocity, spring.velocityY));
-                spring.velocityZ = Math.max(-maxVelocity, Math.min(maxVelocity, spring.velocityZ));
-                
-                // Apply velocities to rotation
-                bone.rotation.x += spring.velocityX * deltaTime;
-                bone.rotation.y += spring.velocityY * deltaTime;
-                bone.rotation.z += spring.velocityZ * deltaTime;
-                
-                // Apply gravity effect based on hierarchy - deeper bones droop more
-                // Check if the bone should be affected by gravity
-                if (spring.hierarchyFactor < 0.8) { // Only affect non-root bones
-                    // Get global gravity value from rigOptions if available, or use a reasonable default
-                    const worldGravity = (rigOptions && rigOptions.worldGravity !== undefined) ? 
-                        rigOptions.worldGravity : 1.0;
-                    
-                    // Simple gravity effect that pulls parts downward around local X axis
-                    // This creates a more natural ragdoll droop effect
-                    // The gravity influence is scaled by hierarchy and the global gravity value
-                    const gravityInfluence = (1 - spring.hierarchyFactor) * 0.5 * Math.abs(worldGravity);
-                    
-                    // Apply gravity with correct direction (positive or negative based on worldGravity sign)
-                    const gravityDirection = worldGravity >= 0 ? 1 : -1;
-                    bone.rotation.x += gravityInfluence * deltaTime * 2.0 * gravityDirection;
-                }
-                
-                // Store current state for next update
-                spring.lastPosition.copy(bone.position);
-                spring.lastTime = now;
-                spring.lastRotation.copy(bone.rotation);
-                
-                // If this bone has children, ensure the children also update their constraints
-                // This ensures force propagation through the hierarchy
-                if (bone.children && bone.children.length > 0) {
-                    for (let i = 0; i < bone.children.length; i++) {
-                        const child = bone.children[i];
-                        if (child.isBone && child.userData.constraints && 
-                            child.userData.constraints.type === 'spring') {
-                            // Force immediate update of child springs for proper cascade effect
-                            enforceJointConstraints(child);
-                        }
-                    }
-                }
-            }
-            break;
-            
-        case 'none':
-        default:
-            // No constraints (default) - unrestricted rotation
-            // No need to do anything special here, allow free movement
-            break;
-    }
-}
-
-/**
- * Analyze the rig data in a GLTF model
- * @param {Object} gltf - The loaded GLTF model data
- * @returns {Object} Analyzed rig details
- */
-export function analyzeGltfModel(gltf) {
-    console.log('analyzeGltfModel called with:', gltf);
-    
-    if (!gltf || !gltf.scene) {
-        console.error('Invalid GLTF model:', gltf);
-        return null;
-    }
-    
-    console.log('Analyzing GLTF model for rig information...');
-    console.log('GLTF scene structure:', gltf.scene);
-    
-    const rawDetails = {
-        bones: [],
-        rigs: [],
-        roots: [],
-        controls: [],
-        joints: [],
-        constraints: [] // Add a new array to track constraints
-    };
-    
-    // Extract scene information
-    const scene = gltf.scene;
-    
-    // Helper function to traverse the scene
-    const traverseNode = (node, parentType = null) => {
-        console.log('Traversing node:', node.name, 'type:', node.type, 'isBone:', node.isBone);
-        
-        // Look for joint constraints in this node
-        const constraints = parseJointConstraints(node);
-        if (constraints) {
-            console.log(`Found constraints for ${node.name}:`, constraints);
-            rawDetails.constraints.push({
-                nodeName: node.name,
-                nodeType: node.type,
-                constraintType: constraints.type,
-                constraintData: constraints
-            });
-        }
-        
-        // Check if the node is a bone
-        if (node.isBone || node.name.toLowerCase().includes('bone')) {
-            console.log('Found bone:', node.name);
-            rawDetails.bones.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null,
-                rotation: node.rotation ? [node.rotation.x, node.rotation.y, node.rotation.z] : null,
-                parentName: parentType === 'bone' ? node.parent.name : null,
-                constraintType: constraints ? constraints.type : 'none' // Add constraint type
-            });
-            parentType = 'bone';
-        }
-        
-        // Check if the node is a rig
-        if (node.name.toLowerCase().includes('rig') || node.name.toLowerCase().includes('armature')) {
-            console.log('Found rig:', node.name);
-            rawDetails.rigs.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null,
-                childCount: node.children ? node.children.length : 0
-            });
-            parentType = 'rig';
-        }
-        
-        // Check if the node is a root
-        if (node.name.toLowerCase().includes('root')) {
-            console.log('Found root:', node.name);
-            rawDetails.roots.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null
-            });
-            parentType = 'root';
-        }
-        
-        // Check if the node is a control/handle
-        if (node.name.toLowerCase().includes('control') || 
-            node.name.toLowerCase().includes('ctrl') || 
-            node.name.toLowerCase().includes('handle')) {
-            console.log('Found control:', node.name);
-            rawDetails.controls.push({
-                name: node.name,
-                position: node.position ? [node.position.x, node.position.y, node.position.z] : null,
-                type: node.name.toLowerCase().includes('control') ? 'control' : 
-                      node.name.toLowerCase().includes('ctrl') ? 'ctrl' : 'handle'
-            });
-            parentType = 'control';
-        }
-        
-        // Traverse children
-        if (node.children) {
-            node.children.forEach(child => traverseNode(child, parentType));
-        }
-    };
-    
-    // Start traversal from the scene
-    console.log('Starting scene traversal');
-    scene.traverse(node => traverseNode(node));
-    
-    // Process raw details to deduplicate items
-    const result = {
-        bones: deduplicateItems(rawDetails.bones),
-        rigs: deduplicateItems(rawDetails.rigs),
-        roots: deduplicateItems(rawDetails.roots),
-        controls: deduplicateItems(rawDetails.controls),
-        joints: deduplicateItems(rawDetails.joints),
-        constraints: rawDetails.constraints // Don't deduplicate constraints
-    };
-    
-    console.log('Rig analysis results:', result);
-    return result;
-}
-
-/**
- * Deduplicate items by name, counting duplicates
- * @param {Array} items - Array of items to deduplicate
- * @returns {Array} Deduplicated items with count property
- */
-export function deduplicateItems(items) {
-    const uniqueItems = new Map();
-    
-    items.forEach(item => {
-        const key = item.name;
-        if (uniqueItems.has(key)) {
-            const existingItem = uniqueItems.get(key);
-            existingItem.count = (existingItem.count || 1) + 1;
-        } else {
-            uniqueItems.set(key, {...item, count: 1});
-        }
-    });
-    
-    return Array.from(uniqueItems.values());
-}
+import { createLabels } from './rig-label-factory';
+import { addControlHandleToFurthestBone, applyJointConstraints } from './rig-constraint-manager';
+import { parseJointConstraints } from '../data/glb-classifier';
 
 /**
  * Create a bone joint with consistent styling and properties
@@ -1246,50 +233,6 @@ export function createBoneUpdateFunction(boneGroup) {
             }
         }
     };
-}
-
-/**
- * Add a control handle to the furthest bone
- * @param {Object} bone - The bone to add the handle to
- * @param {Object} scene - The Three.js scene
- * @param {Number} modelScale - Scale factor for the handle size
- */
-export function addControlHandleToFurthestBone(bone, scene, modelScale) {
-    const handleSize = modelScale * 2.6; // Double the size (from 1.3 to 2.6)
-    const geometry = new THREE.SphereGeometry(handleSize, 16, 16);
-    const material = new THREE.MeshBasicMaterial({
-        color: rigOptions.normalColor, // Use color from rigOptions
-        transparent: true,
-        opacity: 0.7,
-        wireframe: false
-    });
-    // Set the furthest bone handle
-    setFurthestBoneHandle(new THREE.Mesh(geometry, material), "FurthestBoneHandle", scene, bone);
-    console.log('Added control handle to furthest bone:', bone.name);
-}
-
-/**
- * Create both joint and bone labels and set their visibility
- * @param {Object} scene - The Three.js scene
- */
-export function createLabels(scene) {
-    // Create joint labels
-    console.log('Setting up joint labels');
-    createJointLabels(scene);
-    
-    // Check if joint labels should be visible based on option
-    if (!rigOptions.showJointLabels) {
-        hideJointLabels();
-    }
-    
-    // Create bone labels
-    console.log('Setting up bone labels');
-    createBoneLabels(scene);
-    
-    // Check if bone labels should be visible based on option
-    if (!rigOptions.showBoneLabels) {
-        hideBoneLabels();
-    }
 }
 
 /**
